requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.1.2
paypalrestsdk==1.13.1
PyJWT==2.8.0
psycopg2-binary
sqlalchemy
gunicorn==20.1.0